% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDm6A_application.R
\name{TDm6A_application}
\alias{TDm6A_application}
\title{<TDm6A_application>}
\usage{
TDm6A_application(faFile_Input, plot_Output, csvFile_Output, cellType)
}
\arguments{
\item{faFile_Input}{full path of the input fasta file storing the ID and sequence information}

\item{plot_Output}{full path of the output .png file to show the predicted m6A sites on the given sequence}

\item{csvFile_Output}{full path of the output .csv file to store the predicted results}

\item{cellType}{a string to be "A549", or "CD8T", or "HEK293"}
}
\description{
This function is used to apply TDm6A models to predict m6A sites on a given nucleotide sequence
}
