% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TDm6A_test.R
\name{TDm6A_test}
\alias{TDm6A_test}
\title{<TDm6A_test>}
\usage{
TDm6A_test(x_test, y_test, cellType)
}
\arguments{
\item{x_test}{a 3D array}

\item{y_test}{a 2D array}

\item{cellType}{a string to be "A549", or "CD8T", or "HEK293"}
}
\description{
This function test TDm6A models with test case
}
